Version 0.4.5
  Always clear previously overlayed highlighting:
   http://github.com/xolox/vim-lua-inspect/commit/62942b011979937fba84ea4795f7a5c593323f31

  Bug fix for `E776: No location list':
   http://github.com/xolox/vim-lua-inspect/commit/47b25952b638bb02f69c7fb479cbb2d57c806a94

  Make argument count warnings highlight only function names:
   http://github.com/xolox/vim-lua-inspect/commit/22b75bcf657d4f55e48cce3801613ee3e3086af1

Lots of miscellaneous changes (see http://github.com/xolox/vim-lua-inspect/commits/) to make the plug-in compatible with the newest release of LuaInspect, which itself includes bug fixes that fix bugs in this plug-in (e.g. when you add some empty lines to the start of a Lua buffer the highlighting would break, this turned out to be a bug in the Metalua lexer).